split up grid merging and rest of .cpp, push content printer to cpp
