+++ Core/include/Acts/Geometry/GridPortalLink.hpp
@@ -92,7 +92,7 @@ class GridPortalLink : public PortalLinkBase {
                              GridPortalLink& merged, BinningValue direction,
                              const Logger& logger);
 
-  virtual void prettyPrint(std::ostream& os) const = 0;
+  void printContents(std::ostream& os) const;
 
  protected:
   void checkConsistency(const CylinderSurface& cyl) const;
@@ -188,77 +188,6 @@ class GridPortalLinkT final : public GridPortalLink {
     }
   }
 
-  void prettyPrint(std::ostream& os) const final {
-    // @TODO: Possibly move this to .cpp file
-    std::cout << "----- GRID " << GridType::DIM << "d -----" << std::endl;
-    std::cout << grid() << " along " << direction() << std::endl;
-
-    std::string loc0;
-    std::string loc1;
-
-    bool flipped = false;
-
-    if (surface().type() == Surface::Cylinder) {
-      loc0 = "rPhi";
-      loc1 = "z";
-      flipped = direction() != BinningValue::binRPhi;
-    } else if (surface().type() == Surface::Disc) {
-      loc0 = "r";
-      loc1 = "phi";
-      flipped = direction() != BinningValue::binR;
-    } else if (surface().type() == Surface::Plane) {
-      loc0 = "x";
-      loc1 = "y";
-      flipped = direction() != BinningValue::binX;
-    } else {
-      throw std::invalid_argument{"Unsupported surface type"};
-    }
-
-    if constexpr (GridType::DIM == 1) {
-      auto loc = grid().numLocalBins();
-
-      if (flipped) {
-        os << std::format("{} >       i=0 ", loc1);
-        for (std::size_t i = 1; i <= loc[0] + 1; i++) {
-          os << std::format("{:>11} ", std::format("i={}", i));
-        }
-        os << std::endl;
-
-        os << "    ";
-        for (std::size_t i = 0; i <= loc[0] + 1; i++) {
-          const void* v = grid().atLocalBins({i});
-          os << std::format("{:11}", v) << " ";
-        }
-        os << std::endl;
-
-      } else {
-        os << std::format("v {}", loc0) << std::endl;
-        for (std::size_t i = 0; i <= loc[0] + 1; i++) {
-          os << "i=" << i << " ";
-          const void* v = grid().atLocalBins({i});
-          os << std::format("{:11}", v) << " ";
-          os << std::endl;
-        }
-      }
-
-    } else {
-      auto loc = grid().numLocalBins();
-      os << std::format("v {}|{} >   j=0 ", loc0, loc1, 0);
-      for (std::size_t j = 1; j <= loc[1] + 1; j++) {
-        os << std::format("{:>11} ", std::format("j={}", j));
-      }
-      os << std::endl;
-      for (std::size_t i = 0; i <= loc[0] + 1; i++) {
-        os << "i=" << i << " ";
-        for (std::size_t j = 0; j <= loc[1] + 1; j++) {
-          const void* v = grid().atLocalBins({i, j});
-          os << std::format("{:11}", v) << " ";
-        }
-        os << std::endl;
-      }
-    }
-  }
-
  protected:
   std::vector<std::size_t> numLocalBins() const final {
     typename GridType::index_t idx = m_grid.numLocalBins();
@@ -292,4 +221,5 @@ class GridPortalLinkT final : public GridPortalLink {
  private:
   GridType m_grid;
 };
+
 }  // namespace Acts
